+++ packages/pudu/v0.1/src/Pudu/Eval/Operator/Access.hs
@@ -232,6 +232,9 @@ stringMethods =
   , ("chars", StringChars)
   , ("lines", StringLines)
   , ("reverse", StringReverse)
+  , ("toInt", StringToInt)
+  , ("toFloat", StringToFloat)
+  , ("toDecimal", StringToDecimal)
   ]
 
 {-| Method names paired with their method tags. -}